+++ releasenotes.md
@@ -1,6 +1,5 @@
-Whats New in v8.6
+Whats New in v8.7
 --------------------------
-- Restructure formula settings page 
-- Fix an issue preventing formula to display
+- If the materila quantity exceeds the amount in stock, when making a formula, PV will add the material and update material's quantity down to its maximum available, instead if returning an error
 - Various bug fixes and improvements
 - For full details please refer to the CHANGELOG